Make it so that comets don't move during pause.
Because comet movement is based on FC_time_elapsed, and that is based on when
the last FC_frame_begin call was made, when you pause the comets game comets
move to where they would have been if you hadn't paused the game on the next
while loop iteraiton.  This closes the current frame at the begining of a
pause and restarts it afterwards.